// ClientApp/src/app/models/transfer-flat.ts
export interface ITransferFlat {
    id: string
    destination: string
    route: string
    customer: string
    pickupPoint: string
    time: string
    adults: number
    kids: number
    free: number
    port: string
    driver: string
    dateIn: string
    totalPersons: number
    userName: string
    remarks: string
}